last update 5/2024

+++ contract/script/deploySabreDAOEngine.s.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
+
 import {Script} from "../lib/forge-std/src/Script.sol";
 import {console} from "../lib/forge-std/src/console.sol";
 import {helperConfig} from "../script/helperConfig.s.sol";
@@ -11,28 +12,23 @@ import {S_vault} from "../src/S_vault.sol";
 contract deploySabreDAOEngine is Script {
     address[] public proposer;
 
-
-    function run () external returns (SabreDAO, helperConfig, SabreDAOEngine, S_vault, TimeLock) {
+    function run() external returns (SabreDAO, helperConfig, SabreDAOEngine, S_vault, TimeLock) {
         helperConfig HelperConfig = new helperConfig();
-        (uint256 _proposalFee,
-        uint256 _votingFee,
-        uint256 _timePoint,
-        uint _proposalID,
-        // address sabreDAOAddress;
-        uint deployer_Key
+        (
+            uint256 _proposalFee,
+            uint256 _votingFee,
+            uint256 _timePoint,
+            uint256 _proposalID,
+            // address sabreDAOAddress;
+            uint256 deployer_Key
         ) = HelperConfig.activeNetworkConfig();
         vm.startBroadcast(deployer_Key);
 
-        SabreDAO SBRToken = new  SabreDAO(msg.sender);
+        SabreDAO SBRToken = new SabreDAO(msg.sender);
         S_vault SBRVault = new S_vault(address(SBRToken));
 
-        SabreDAOEngine SBREngine = new SabreDAOEngine(
-            _proposalFee,
-            _votingFee,
-            _timePoint,
-            _proposalID,
-            address(SBRToken)    
-        );
+        SabreDAOEngine SBREngine =
+            new SabreDAOEngine(_proposalFee, _votingFee, _timePoint, _proposalID, address(SBRToken));
         TimeLock timeLock = new TimeLock(30, proposer, proposer);
 
         // SBRToken.mintToken();
@@ -40,5 +36,4 @@ contract deploySabreDAOEngine is Script {
         vm.stopBroadcast();
         return (SBRToken, HelperConfig, SBREngine, SBRVault, timeLock);
     }
-    
 }
